WIP: separate reveal fixture and set up unnamed accounts
